#2898 Templates (Word/Excel)

+++ PrimeApps.Console/wwwroot/view/app/templates/exceltemplates/excelTemplatesController.js
@@ -13,6 +13,14 @@ angular.module('primeapps')
             $rootScope.breadcrumblist[1].link = '#/org/' + $rootScope.currentOrganization.id + '/app/' + $rootScope.appId + '/overview';
             $rootScope.breadcrumblist[2].title = 'Excel Templates';
 
+            $scope.generator = function (limit) {
+                $scope.placeholderArray = [];
+                for (var i = 0; i < limit; i++) {
+                    $scope.placeholderArray[i] = i;
+                }
+            };
+            $scope.generator(10);
+
             $scope.loading = true;
 
             $scope.requestModel = {
@@ -32,6 +40,7 @@ angular.module('primeapps')
                     template.module = $filter('filter')($scope.$parent.modules, { name: template.module }, true)[0];
                 });
                 $scope.templates = templates;
+                $scope.templatesState = templates;
 
             }).finally(function () {
                 $scope.loading = false;
@@ -49,6 +58,7 @@ angular.module('primeapps')
                         template.module = $filter('filter')($scope.$parent.modules, { name: template.module }, true)[0];
                     });
                     $scope.templates = templates;
+                    $scope.templatesState = templates;
 
                 }).finally(function () {
                     $scope.loading = false;
@@ -223,9 +233,28 @@ angular.module('primeapps')
             };
 
             $scope.delete = function (id) {
-                ExcelTemplatesService.delete(id).then(function () {
-                    swal($filter('translate')('Setup.Templates.DeleteSuccess' | translate), "", "success");
-                });
+                const willDelete =
+                    swal({
+                        title: "Are you sure?",
+                        text: "Are you sure that you want to delete this dependency ?",
+                        icon: "warning",
+                        buttons: ['Cancel', 'Okey'],
+                        dangerMode: true
+                    }).then(function (value) {
+                        if (value) {
+                            ExcelTemplatesService.delete(id).then(function () {
+                                $scope.changePage(1);
+                                swal($filter('translate')('Setup.Templates.DeleteSuccess' | translate), "", "success");
+                            }).catch(function () {
+                                $scope.templates = $scope.templatesState;
+
+                                if ($scope.addNewExcelTemplateFormModal) {
+                                    $scope.addNewExcelTemplateFormModal.hide();
+                                    $scope.saving = false;
+                                }
+                            });
+                        }
+                    });
             };
 
         }
